requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.6.1
supabase==2.9.1
python-jose[cryptography]==3.3.0
cryptography==43.0.3
google-generativeai==0.8.3
python-docx==1.1.2
httpx==0.27.2
python-dotenv==1.0.1
ruff==0.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
